+++ cli/docs/changelog.md
@@ -5,6 +5,9 @@
 ### Features
 * `kurtosis gateway` now connects to user services running in cluster
 
+### Fixes
+* Fixed a bug where the gateway wasn't rewriting user service requests with the local ports
+
 # 0.18.4
 ### Features
 * Added files artifact expansion in Kubernetes